long overdue update

+++ laravel-vue/ticket-system/app/Http/Controllers/CategoryController.php
@@ -9,6 +9,7 @@ use App\Models\Category;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Http\Resources\Json\ResourceCollection;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Illuminate\Http\Exceptions\HttpResponseException;
 
 class CategoryController extends Controller
 {
@@ -21,16 +22,10 @@ class CategoryController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     */
-    public function show(Category $category) : JsonResource {
-        return new CategoryResource($category);
-    }
-
-    /**
      * Store a newly created resource in storage.
      */
     public function store(StoreCategoryRequest $request) : JsonResource {
+        $this->requiresAdmin();
         $category = Category::create($request->validated());
         return new CategoryResource($category);
     }
@@ -39,6 +34,7 @@ class CategoryController extends Controller
      * Update the specified resource in storage.
      */
     public function update(UpdateCategoryRequest $request, Category $category) : JsonResource {
+        $this->requiresAdmin();
         $category->update($request->validated());
         return new CategoryResource($category);
     }
@@ -47,7 +43,16 @@ class CategoryController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(Category $category) : JsonResponse {
+        $this->requiresAdmin();
+
+        if ($category->tickets()->exists()) {
+            throw new HttpResponseException(response()->json([
+                'message' => 'Category cannot be deleted because they are still being used by tickets.',
+                'errors' => []
+            ], 422));    
+        }
+
         $category->delete();
-        return response()->json(['message' => 'resource was deleted successfully']);
+        return response()->json(['message' => 'resource was deleted successfully']); 
     }
 }
